Aquatic Paddock 3rdt Test Passing

+++ src/main/java/paddocks/AquaticPaddock.java
@@ -2,7 +2,7 @@ package paddocks;
 
 public class AquaticPaddock extends Paddocks{
 
-    public AquaticPaddock(String name, String type, double size, int defenseValue){
+    public AquaticPaddock(String name, String type, int size, int defenseValue){
         super(name, type, size, defenseValue);
     }
